// src/DeviceManufacturer.d.ts
declare enum DeviceManufacturer {
    SEQUENTIAL_CIRCUITS = 1,
    BIG_BRIAR = 2,
    OCTAVE_PLATEAU = 3,
    MOOG = 4,
    PASSPORT_DESIGNS = 5,
    LEXICON = 6,
    KURZWEIL = 7,
    FENDER = 8,
    GULBRANSEN = 9,
    DELTA_LABS = 10,
    SOUND_COMP = 11,
    GENERAL_ELECTRO = 12,
    TECHMAR = 13,
    MATTHEWS_RESEARCH = 14,
    OBERHEIM = 16,
    PAIA = 17,
    SIMMONS = 18,
    GENTLE_ELECTRIC = 19,
    FAIRLIGHT = 20,
    JL_COOPER = 21,
    LOWERY = 22,
    LIN = 23,
    EMU = 24,
    PEAVEY = 27,
    BON_TEMPI = 32,
    SIEL = 33,
    SYNTHEAXE = 35,
    HOHNER = 36,
    CRUMAR = 37,
    SOLTON = 38,
    JELLINGHOUS_MS = 39,
    CTS = 40,
    PPG = 41,
    ELKA = 47,
    KAWAI = 64,
    ROLAND = 65,
    KORG = 66,
    YAMAHA = 67,
    CASIO = 68,
    AKAI = 69
}
export default DeviceManufacturer;