// tests/EventByteArrays.d.ts
declare const ByteArrays: {
    TEXT: number[];
    COPYRIGHT: number[];
    TRACK_NAME: number[];
    INSTRUMENT_NAME: number[];
    LYRIC: number[];
    MARKER: number[];
    CUE_POINT: number[];
    SET_TEMPO: number[];
    SMTPE_OFFSET: number[];
    SEQUENCE_NUMBER: number[];
    END_OF_TRACK: number[];
    CHANNEL_PREFIX: number[];
    PORT_PREFIX: number[];
    KEY_SIGNATURE: number[];
    TIME_SIGNATURE: number[];
    SEQUENCER_SPECIFIC: number[];
    SYSEX: number[];
    NOTE_ON: number[];
    NOTE_OFF: number[];
    AFTERTOUCH: number[];
    CONTROLLER: number[];
    PROGRAM_CHANGE: number[];
    CHANNEL_AFTERTOUCH: number[];
    PITCH_WHEEL: number[];
    RUNNING_C_MAJOR_TRIAD: number[];
    INVALID_TEXT: number[];
    INVALID_META_EVENT_TYPE: number[];
};
export default ByteArrays;